refactor(map): MapProvider entities 계층으로 이동
