+++ src/app/routes/AppRouter.tsx
@@ -8,7 +8,7 @@ import FullPageLayout from '@/layouts/FullPageLayout';
 
 // Lazy load pages
 const Home = lazy(() => import('@/pages/Home'));
-const Map = lazy(() => import('@/pages/Map/index'));
+const Map = lazy(() => import('@/pages/Map'));
 const NotFound = lazy(() => import('@/pages/NotFound'));
 const Login = lazy(() => import('@/pages/Login'));
 const SignUp = lazy(() => import('@/pages/SignUp'));
