+++ src/entities/map/model/MapContext.tsx
@@ -1,7 +1,7 @@
-import { createContext, useContext, type ReactNode } from 'react';
 import useMap from '@/features/Map/model/hooks/useMap';
 import type { MapType } from '@/features/Map/model/types';
-import { Map } from 'ol';
+import type { Map } from 'ol';
+import { createContext, type PropsWithChildren } from 'react';
 
 interface MapContextType {
   mapRef: React.RefObject<HTMLDivElement | null>;
@@ -15,16 +15,10 @@ interface MapContextType {
 
 const MapContext = createContext<MapContextType | null>(null);
 
-export function MapProvider({ children }: { children: ReactNode }) {
+export function MapProvider({ children }: PropsWithChildren) {
   const mapData = useMap();
 
   return <MapContext.Provider value={mapData}>{children}</MapContext.Provider>;
 }
 
-export function useMapContext() {
-  const context = useContext(MapContext);
-  if (!context) {
-    throw new Error('변경 중 오류가 발생했습니다.');
-  }
-  return context;
-}
+export default MapContext;
